requirements.txt
streamlit
crewai
openai
langchain
python-dotenv
pypdf
tqdm
langchain-community
faiss-cpu
langchain-openai
